FEAT: add `tool.nbqa.addopts` `ruff` setting
